Revert "add icons to repo settings tabs and move actions"
This reverts commit 95a75295b9a47512c57c3b984334ac4d02a4ee49.
