+++ web/src/views/repo/settings/General.vue
@@ -167,59 +167,13 @@
         :text="$t('repo.settings.general.save')"
       />
     </form>
-
-    <div class="dark:border-wp-background-100 my-4 flex flex-col justify-center border-b pb-4">
-      <h1 class="text-wp-text-100 flex items-center gap-1 text-xl">
-        {{ $t('repo.settings.actions.actions') }}
-      </h1>
-    </div>
-
-    <div class="flex flex-wrap items-center">
-      <Button
-        class="my-1 mr-4"
-        color="blue"
-        start-icon="heal"
-        :is-loading="isRepairingRepo"
-        :text="$t('repo.settings.actions.repair.repair')"
-        @click="repairRepo"
-      />
-
-      <Button
-        v-if="isActive"
-        color="blue"
-        class="my-1 mr-4"
-        start-icon="turn-off"
-        :is-loading="isDeactivatingRepo"
-        :text="$t('repo.settings.actions.disable.disable')"
-        @click="deactivateRepo"
-      />
-      <Button
-        v-else
-        class="my-1 mr-4"
-        color="blue"
-        start-icon="turn-off"
-        :is-loading="isActivatingRepo"
-        :text="$t('repo.settings.actions.enable.enable')"
-        @click="activateRepo"
-      />
-
-      <Button
-        class="my-1 mr-4"
-        color="red"
-        start-icon="trash"
-        :is-loading="isDeletingRepo"
-        :text="$t('repo.settings.actions.delete.delete')"
-        @click="deleteRepo"
-      />
-    </div>
   </Settings>
 </template>
 
 <script lang="ts" setup>
-import { computed, inject, onMounted, ref } from 'vue';
+import { inject, onMounted, ref } from 'vue';
 import type { Ref } from 'vue';
 import { useI18n } from 'vue-i18n';
-import { useRouter } from 'vue-router';
 
 import Button from '~/components/atomic/Button.vue';
 import Checkbox from '~/components/form/Checkbox.vue';
@@ -239,61 +193,12 @@ import type { Repo, RepoSettings } from '~/lib/api/types';
 import { useRepoStore } from '~/store/repos';
 
 const apiClient = useApiClient();
-const router = useRouter();
 const notifications = useNotifications();
+const { user } = useAuthentication();
+const repoStore = useRepoStore();
 const i18n = useI18n();
 
 const repo = inject<Ref<Repo>>('repo');
-
-const { doSubmit: repairRepo, isLoading: isRepairingRepo } = useAsyncAction(async () => {
-  if (!repo) {
-    throw new Error('Unexpected: Repo should be set');
-  }
-
-  await apiClient.repairRepo(repo.value.id);
-  notifications.notify({ title: i18n.t('repo.settings.actions.repair.success'), type: 'success' });
-});
-
-const { doSubmit: deleteRepo, isLoading: isDeletingRepo } = useAsyncAction(async () => {
-  if (!repo) {
-    throw new Error('Unexpected: Repo should be set');
-  }
-
-  // TODO: use proper dialog
-  // eslint-disable-next-line no-alert
-  if (!confirm(i18n.t('repo.settings.actions.delete.confirm'))) {
-    return;
-  }
-
-  await apiClient.deleteRepo(repo.value.id);
-  notifications.notify({ title: i18n.t('repo.settings.actions.delete.success'), type: 'success' });
-  await router.replace({ name: 'repos' });
-});
-
-const { doSubmit: activateRepo, isLoading: isActivatingRepo } = useAsyncAction(async () => {
-  if (!repo) {
-    throw new Error('Unexpected: Repo should be set');
-  }
-
-  await apiClient.activateRepo(repo.value.forge_remote_id);
-  notifications.notify({ title: i18n.t('repo.settings.actions.enable.success'), type: 'success' });
-});
-
-const { doSubmit: deactivateRepo, isLoading: isDeactivatingRepo } = useAsyncAction(async () => {
-  if (!repo) {
-    throw new Error('Unexpected: Repo should be set');
-  }
-
-  await apiClient.deleteRepo(repo.value.id, false);
-  notifications.notify({ title: i18n.t('repo.settings.actions.disable.success'), type: 'success' });
-  await router.replace({ name: 'repos' });
-});
-
-const isActive = computed(() => repo?.value.active);
-
-const { user } = useAuthentication();
-const repoStore = useRepoStore();
-
 const repoSettings = ref<RepoSettings>();
 
 function loadRepoSettings() {
